✨ Implementation of process to scan all directories at startup and generate stories file
