+++ packages/auto-story-generator/.vscode/settings.json
@@ -3,7 +3,8 @@
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.tabSize": 2,
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": "explicit"
+    "source.fixAll.eslint": "explicit",
+    "source.addMissingImports": "explicit"
   },
   "files.eol": "\n",
   "typescript.preferences.importModuleSpecifier": "non-relative"
