+++ packages/auto-story-generator/src/core/genStoryFile.ts
