+++ packages/auto-story-generator/src/core/getComponentInfo.ts
