+++ packages/auto-story-generator/src/presets/lit/genLitStoryFile.ts
@@ -1,8 +1,8 @@
 import { kebabCase, pascalCase } from "scule";
 
+import { genStoryFile } from "~/src/core/genStoryFile";
 import { getLitPropTypes } from "~/src/presets/lit/getLitPropTypes";
 import { GenStoryFileOptions } from "~/src/types/GenStoryFileType";
-import { genStoryFile } from "~/src/utils/genStoryFile";
 import { throwErr } from "~/src/utils/throwError";
 
 export const genLitStoryFile = async ({
